+++ apps/flutter_sample_cocoapods/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:customer_io/customer_io.dart';
+import 'package:zixflow/zixflow.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';