+++ apps/flutter_sample_spm/lib/src/random.dart
@@ -21,7 +21,7 @@ class RandomValues {
       result += _emailUsernameWhitelistChars[random.nextInt(charsLength)];
     }
 
-    return "$result@customer.io";
+    return "$result@zixflow.com";
   }
 
   MapEntry<String, Map<String, Object>?> trackingEvent() {