+++ apps/flutter_sample_spm/lib/src/screens/inline_messages.dart
@@ -1,6 +1,6 @@
-import 'package:customer_io/customer_io.dart';
-import 'package:customer_io/customer_io_widgets.dart';
-import 'package:customer_io/messaging_in_app/inline_in_app_message_view.dart';
+import 'package:zixflow/zixflow.dart';
+import 'package:zixflow/customer_io_widgets.dart';
+import 'package:zixflow/messaging_in_app/inline_in_app_message_view.dart';
 import 'package:flutter/material.dart';
 
 import '../components/container.dart';