+++ apps/flutter_sample_spm/lib/src/screens/location.dart
@@ -1,4 +1,4 @@
-import 'package:customer_io/customer_io.dart';
+import 'package:zixflow/zixflow.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart' show MethodChannel;