+++ test/messaging_in_app/animation_test.dart
@@ -1,4 +1,4 @@
-import 'package:customer_io/customer_io_widgets.dart';
+import 'package:zixflow/customer_io_widgets.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';