+++ test/messaging_in_app/test_helpers.dart
@@ -1,4 +1,4 @@
-import 'package:customer_io/messaging_in_app/inline_in_app_message_view.dart';
+import 'package:zixflow/messaging_in_app/inline_in_app_message_view.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';